backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
linkedin-api==2.0.3
indeed-python==0.0.6
python-jose==3.3.0